Fix mix-in-length issue
