+++ eth2/utils/tree_hash/src/cached_tree_hash/impls.rs
@@ -3,13 +3,16 @@ use super::*;
 mod vec;
 
 impl CachedTreeHashSubTree<u64> for u64 {
-    fn new_tree_hash_cache(&self, depth: usize) -> Result<TreeHashCache, Error> {
+    fn new_tree_hash_cache(&self, _depth: usize) -> Result<TreeHashCache, Error> {
         Ok(TreeHashCache::from_bytes(
             merkleize(self.to_le_bytes().to_vec()),
             false,
-            // self.tree_hash_cache_overlay(0, depth)?,
             None,
         )?)
+    }
+
+    fn num_tree_hash_cache_chunks(&self) -> usize {
+        1
     }
 
     fn tree_hash_cache_overlay(
@@ -17,7 +20,41 @@ impl CachedTreeHashSubTree<u64> for u64 {
         chunk_offset: usize,
         depth: usize,
     ) -> Result<BTreeOverlay, Error> {
-        BTreeOverlay::from_lengths(chunk_offset, 1, depth, vec![1])
+        panic!("Basic should not produce overlay");
+        // BTreeOverlay::from_lengths(chunk_offset, 1, depth, vec![1])
+    }
+
+    fn update_tree_hash_cache(&self, cache: &mut TreeHashCache) -> Result<(), Error> {
+        let leaf = merkleize(self.to_le_bytes().to_vec());
+        cache.maybe_update_chunk(cache.chunk_index, &leaf)?;
+
+        cache.chunk_index += 1;
+        // cache.overlay_index += 1;
+
+        Ok(())
+    }
+}
+
+impl CachedTreeHashSubTree<usize> for usize {
+    fn new_tree_hash_cache(&self, _depth: usize) -> Result<TreeHashCache, Error> {
+        Ok(TreeHashCache::from_bytes(
+            merkleize(self.to_le_bytes().to_vec()),
+            false,
+            None,
+        )?)
+    }
+
+    fn num_tree_hash_cache_chunks(&self) -> usize {
+        1
+    }
+
+    fn tree_hash_cache_overlay(
+        &self,
+        chunk_offset: usize,
+        depth: usize,
+    ) -> Result<BTreeOverlay, Error> {
+        panic!("Basic should not produce overlay");
+        // BTreeOverlay::from_lengths(chunk_offset, 1, depth, vec![1])
     }
 
     fn update_tree_hash_cache(&self, cache: &mut TreeHashCache) -> Result<(), Error> {
